+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 3.7.4 (22.06.2018)
 * Implemented localization support via .properties files - more info later
+* Added whitelisted commands configurable in config.yml
+* Now admin commands won't be blocked via blocked commands in game
 
 ### 3.7.3 Release (14.06.2018)
 * Fixed boss bar startup errors